+++ src/features/card/actions/index.ts
@@ -1,0 +1,2 @@
+export * from './ui';
+export * as cardActionsModel from './model';
