+++ src/features/card/actions/ui/index.ts
@@ -1,0 +1,6 @@
+// FIXME: maybe decompose later (to separate features)
+
+export * from './update-changes';
+export * from './reset-changes';
+export * from './edit-title';
+export * from './edit-content';
