+++ src/pages/card/edit/model.ts
@@ -1,5 +1,6 @@
 import { StartParams } from '@box/lib/page-routing';
 import { attach, combine, createEvent, restore, sample } from 'effector-root';
+import { cardActionsModel } from '@box/features/card/actions';
 import { cardModel } from '@box/entities/card';
 import { historyPush } from '@box/entities/navigation';
 
@@ -7,12 +8,10 @@ import { paths } from '../../paths';
 
 // FIXME: dry with pages/card/view
 export const getCardByIdFx = attach({ effect: cardModel.getCardByIdFx });
-export const submitChangesFx = attach({
-  effect: cardModel.draft.submitChangesFx,
-});
 export const pageLoaded = createEvent<StartParams>();
 export const $pagePending = restore(getCardByIdFx.pending.updates, true);
 
+// Подгружаем данные после монтирования страницы
 sample({
   source: pageLoaded,
   fn: ({ params }) => params.cardId,
@@ -34,12 +33,12 @@ export const $pageTitle = combine(
 // Возвращаем на страницу карточки после сохранения/отмены изменений
 sample({
   // FIXME: process response success
-  source: submitChangesFx.done,
+  source: cardActionsModel.submitChangesFx.done,
   fn: ({ params }) => paths.card(params.id),
   target: historyPush,
 });
 sample({
-  source: cardModel.draft.resetChanges,
+  source: cardActionsModel.resetChanges,
   // FIXME: get from store
   fn: (cartId) => paths.card(cartId),
   target: historyPush,
